Assemble messages splitted in the header.
Refactor of MessageAssembler to accomodate this change.

This is kind of a hack, because a PartialMessage can now be either a
partial message or a partial header. If a PartialMessage has

  buffer == NULL,

then it is still waiting for header bytes. For example, a PartialMessage
with:

  bytes_received = 1;
  message_length = 0x??000000;
  buffer == NULL;

contains the first byte of the header in message_length.